bug fix
Changes Made:

I moved the messageText.Replace() operation outside the int.TryParse() calls to avoid repeating the same string manipulation.
I replaced SkipWhile() and TakeWhile() with Where() in order to simplify the code and make it easier to read.
I used Math.Max() and Math.Min() instead of ternary operators to make the code more readable.
I refactored the message string creation to use string concatenation instead of a ternary operator.
I added an if statement to check if the Heart Rate is greater than 0 before adding it to the message. This avoids displaying 0 if there is no Heart Rate data.